ci: remove the TODO of having ansible create the container
Ansible apparently does not have the ability to build containers, and
the Red Hat's suggested way[1] is to call "podman build" on a
Ansible-managed Dockerfile.

That being said, it is possible for Ansible to execute buildah
commands, but one has to wonder why break the symmetry with the
creation of the other containers. Additionally, this would make the
usecase of live-updating a running instance more complicated, as we
would need two codepaths in Ansible for that.

[1] https://www.redhat.com/sysadmin/container-images-ansible