executor/mars: do not taint the DB if lists are in a different order
The list of tags for a machine is not ordered, yet comparing two
dictionaries that reference the same list but in different order
will return that they are different.

To work around this issue, we can use the deepdiff library and ask it
to ignore the order. This also allows us to print the list of changes,
which improves debuggability.