ansible/{salad,executor}: copy packages to install
This copies the respective packages so they are local to the system,
which fixes 'make vivian-provision'.

In the case where the container is being built (i.e. not using
vivian-provision), the copy is basically a no-op since it already exists
in the valve-infra root dir. I think we want to keep this behavior,
because it's possible the user might be iterating over changes to these
packages and want to run provision multiple times. So skipping the copy
if the dest dir already existed (to handle the case where the container
is being built..) would result in changes to the packages not making it
to the VM when running provision multiple times.