ansible/salad: improve restart when config is changed, achieve idempotency
In addition to dealing with idempotency problems in a similar way to the
executor patch, this also properly restarts salad by 1) stopping
executor, 2) restarting salad, 3) starting executor.

Note that handlers that listen to the same thing are executed in the
order they are defined, allowing us to preserve order of operations
without doing anything fancy.