executor/job: expose the machine object to the job description
This is useful for setting the machine's hostname to its MaRS full
name, or many other potential uses.

To make sure that the job won't be able to do something funky with
the original data, we make a copy of the attributes and limit it to
machine attributes and not secrets.