ansible/dashboard/services: allow specifying expected status for service
This allows passing a "name:expected status" to the script, and the
color is selected based on whether the current status matches the
expected status.

This is useful for services that might not be "active" when things are
all OK, e.g. if the service is a oneshot thing like nftables.service. We
don't want to alarm folks by having that be red or yellow, but still
want to make it very clear when the status is not expected (e.g.
"failed")

If an expected status isn't specified, it defaults to "active". The
color for "activating" is still yellow, since I don't think any service
would be expected to be stuck in "activating" and I think it's useful to
show that status as visually different than the others.