telegraf/extra-inputs: capture `scaling_cur_freq`
It was not enabled because of how long it takes to collect the data,
but it also seems to be the only way to get the actual CPU clock on
AMD processors.

On my 5950xt (32 threads), polling goes from 90ms to 480ms, which is well
within the 1s polling period.

Anyway, I doubt anyone will use the current plugin and expect
samples from the different CPUs to be collected at the exact same time,
and this slowness may actually spread the CPU load over a longer period
of time thus reducing the instantenous impact on monitoring.