ansible/chrony: allow client connections from any IP
This removes the hardcoded subnet that chrony is allowed to service NTP
requests on, and configures it to allow requests from *any* IP.
Obviously that's worse for security, so the firewall is relied on to
actually constrain this to the 'private' interface by accepting only
port 123 (NTP) on that iface and dropping connections from any other
interface to that port. Fortunately the firewall was already configured
to do this!

Also see: #60 in the valve-infra gitlab