ansible/network_setup: add entry for ci-gateway to /etc/hosts
This has to be configured each time on boot since podman overwrites /etc/hosts
in the container. Another approach would be to tell podman to not overwrite the
hosts file, but that seemed to be a little more delicate to handle the last
time I tried it.

sed doesn't work here since it tries to overwrite the hosts file, which doesn't
work since it is mounted from podman.

nss doesn't seem to care that 127.0.0.1 is on multiple lines, both localhost
and ci-gateway resolve to 127.0.0.1.