executor: generate dhcp ip range from private interface IP
This is a quick and dirty way to generate the dhcp range, by using the
first 24 bits of the private interface IP and specifying a range from
.10-.100 in that network. If the private interface IP is in that range,
an error is raised. This could be improved later to craft a range that
excludes the private interface IP, but doing that now seemed like a
premature optimization.