makefile: handle case of stopped local registry container
The check here didn't account for cases where the registry container
exists but is *not* started (don't ask me why, maybe because of a sudden
shutdown/reboot or ???).

This would cause the check to always return true (-a will show stopped
containers that exist), and the container would not be started, causing
the build to ultimately fail when it tried to push to the registry.