vivian: don't overwrite config.env on every invocation
This sets the FARM_NAME definition in config.env in a way that doesn't
overwrite the whole file each time, preserving any previously configured
env in the file.

This is an improvement over using something like sed to one-line it,
since it properly handles the case where config.env is an empty file or
missing.