ansible/network_setup: change service type to notify
This script (through find_nic) can take an indeterminate amount of time
to run, making it unreliable with systemd.unit's "before=" directives.
It's unreliable because the service type is "simple" so systemd
considers it started basically as soon as it is executed.

Using systemd's notify interface, we can tell systemd exactly when the
script is done so that things that need to start after it can be started
only after it is done.