executor/gitlab: allow multiple concurrent jobs on the gateway runner
By allowing multiple concurrent jobs to be executed in the gateway
runner, we can run increase the maximum job throughput of the gateway.
This should be very visible on valve-infra pipelines, which all run
on our CI gateways, and thus could have more jobs running in parallel.

We however do not want to enable as many concurrent jobs as we have cpu
cores, as this could potentially lead to widely-different execution
times between runs. Instead, we will limit concurency to cpu_count/8,
which would be equivalent to giving 8 cores per job :) We also make
sure never to set a value of 0, as it would remove any limit.