executor: API REST, add HOST/api/v1/pdus request
The URL http://HOST/api/v1/pdu used to be handled by the MaRS
service (hosted independently its own repo) and got lost when
MaRS was merged in the executor because it wasn't used.
This is not the case anymore as we are trying to make an
interactive dashboard which requires knowledge of the PDUs,
and their ports' states.

So, let's re-introduce this URL, but using the plural form
"pdus" to keep it consistent with the http://HOST/api/v1/machines
request.