executor/mars: only unregister runners when retiring them
This will allow unexposing machines, while still holding on to their
tokens, and thus not having to ask GitLab admins for a valid
registration token every time.

This is the first step towards holding on tokens, and even having a
little stash of them so that adding machines does not always end up
in wasting GitLab admins' time.

Closes: #70