ansible/network_setup: rename private iface and use "private" for bridge
This renames the interface "private" to "privateif" and uses the name
"private" for the bridge name. The reasons for this is to allow existing
apps that bind to "private" to continue functioning, and to avoid having
to add potentially complex config/logic to each one to allow making the
interface name configurable.