vivian: fix setting ssh key id
This conditional test was wrong, it should have been:

    if ! [ -z "$__ssh_id" ]; then

As a result, my key was never used to connect and vivian just sat in a
loop forever (wait_for_ssh) trying, and failing, to connect without -i.

I replaced it with -n because it's nicer, and Shellcheck also recommends
(SC2236) using -n instead of ! -z, though the reasons are purely
stylistic.