executor: REST API, add new endpoint /machine/discover
Add new endpoint /machine/discover that allows interacting with
the ports of a pdu.

The following request:

    curl -X POST HOST/api/v1/machine/discover \
        -H 'Content-Type: application/json' \
        -d '{"port_id": "8", "pdu": "VPDU"}'

will start a discovery process powering the port_id 8 in the pdu VPDU.
If the machine is powered on successfully, it'll add the data of the
on-going discovery to mars.discover_data

If the request is made with a GET method:

    curl -sL HOST/api/v1/machine/discover

the response will give the information of any discovery process ongoing
and the data when it started or will be empty.

Things can go wrong and after powering a port_id from the pdu,
the machine behind it doesn't boot. In that case, a request with
the method DELETE will erase the discovery data:

    curl -X DELETE HOST/api/v1/machine/discover