ansible/minio: create password(s) on first boot
this uses a new oneshot service to generate any passwords for minio, so
that they are no longer stored in the container image.

more passwords, if necessary, can be added to the for loop in
minio_pwgen later on, which is why I went with a loop here even though
there's currently only 1 password/variable.